fix(cron): Add check jobExist before create / delete

+++ apps/back/src/app/cron/cron.service.ts
@@ -16,15 +16,21 @@ export class CronService {
 	}
 
 	addCron(notification: NotificationDto) {
-		const job = new CronJob(notification.cron, async () => {
-			await this.sendNotification(notification);
-		}, null, null, 'Europe/Paris');
-		this.schedulerRegistry.addCronJob(notification.id.toString(), job);
-		job.start();
+		const jobExists = this.schedulerRegistry.doesExist('cron', notification.id.toString());
+		if (!jobExists) {
+			const job = new CronJob(notification.cron, async () => {
+				await this.sendNotification(notification);
+			}, null, null, 'Europe/Paris');
+			this.schedulerRegistry.addCronJob(notification.id.toString(), job);
+			job.start();
+		}
 	}
 
 	deleteCron(notificationId: number) {
-		this.schedulerRegistry.deleteCronJob(notificationId.toString());
+		const jobExists = this.schedulerRegistry.doesExist('cron', notificationId.toString());
+		if (!jobExists) {
+			this.schedulerRegistry.deleteCronJob(notificationId.toString());
+		}
 	}
 
 	async sendNotification(notification: NotificationDto) {
